Config loading and unique attribute rework.

+++ src/main/java/me/spirek/magicalTools/tools/tools/GravityTool.java
@@ -31,7 +31,7 @@ public class GravityTool extends Tool {
         );
 
         addCustomAttribute("gravity",10D);
-        addCustomAttribute("chance",(int)50);
+        addCustomAttribute("chance",50);
     }
 
     public void onAttack(EntityDamageByEntityEvent event){
